+++ css/landing-page.css
@@ -58,9 +58,6 @@ main{
   background-size: cover;
   background-position: right;
 }
-.footer{
-  height: 10%;
-}
 #titulo-landing {
   margin-top: 15vh;
   margin-left: 5vw;
@@ -102,9 +99,9 @@ font-weight: 600;
   font-size: 10px !important;
 } */
 @media screen and (min-width: 576px) {
-  .icons-footer{
+  /* .icons-footer{
     transform: translateX(-7%);
-  }
+  } */
 }
 @media screen and (min-width: 678px) {
   .text-landing{
@@ -119,9 +116,9 @@ font-weight: 600;
     margin-left: 20vw;
     width: 50vw !important;
   }
-  .footer-links{
+  /* .footer-links{
     margin-left: 25% !important;
-  }
+  } */
 }
 @media screen and (min-width: 850px) {
   #titulo-landing {
@@ -172,9 +169,9 @@ font-weight: 600;
   #titulo-landing {
     margin-top: 11vh !important;
   }
-  .container-footer-links{
+  /* .container-footer-links{
     transform: translateX(-5%) !important;
-  }
+  } */
   #main-subtitle{
     margin: 3px 0px !important;
     line-height: 1em;
@@ -371,8 +368,8 @@ font-weight: 600;
   font-weight: 600;
   text-align: start;
 }
-/** Footer */
-.icons-redes{
+/* * Footer */
+/* .icons-redes{
   height: 80px;
   width: 80px;
   font-size: 60px !important;
@@ -382,7 +379,6 @@ font-weight: 600;
   color: var(--color-verde-aguila) !important;
 }
 .footer-links a{
-  /* color: var(--color-blue) !important; */
   color: rgb(67, 125, 159) !important;
 
   font-weight: 750;
@@ -396,31 +392,18 @@ font-weight: 600;
   font-size: 35px !important;
   transform: translateX(-5%);
   text-decoration: none;
-  /* color: var(--color-blue) !important; */
   color: var(--color-verde-aguila) !important;
 }
 .container-footer-bottom{
-  /* color: var(--color-verde-aguila) !important; */
   color: rgb(125, 138, 145) !important;
   font-weight: 700;
 }
 .footer-landing{
-  /* background-color: #10161c; */
   background: linear-gradient(rgba(20, 29, 50) 0%, rgba(20, 29, 50, .8) 100% ) !important;
-}
+} */
 
 /** Final footer  */
-/* --color-cadete: #212738;
---color-verde-a: #0F5257;
---color-lavanda: #6D6875;
---color-gris-topo: #7A7682;
---color-gris-claro: #C8D0CD;
---color-gainsboro: #DCE2E5;
---color-siena-oscuro: #2c0f14;
---color-rojo-toscano: gold;
---color-malva-topo: #926269;
---color-siena-oscuro2: #21090d;
---color-footer: #1B1B1E; */
+
 .preguntas-frecuentes{
   color: var(--color-gris-claro) !important;
   font-style: italic;
@@ -439,13 +422,6 @@ font-weight: 600;
   background: linear-gradient(rgba(20, 29, 50, .5) 0%, rgba(38, 110, 160, 0) 100% );
 }
 
-.footer-landing a:hover{
-  color: var(--color-font-footer-mg) !important;
-  
-}
-.footer-links{
-  margin-left: -20% !important;
-}
