+++ index.html
@@ -18,11 +18,12 @@
     <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet">
     <link href="https://kit-pro.fontawesome.com/releases/v5.15.4/css/pro.min.css" rel="stylesheet">
     <link rel="stylesheet" href="css/landing-page.css" />
+    <link rel="stylesheet" href="css/footer.css">
   </head>
   <body>
     <header>
     </header>
-    <main>      
+    <main>
     </main>
     <footer>
     </footer>
@@ -31,7 +32,9 @@
       integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN"
       crossorigin="anonymous"
     ></script>
-    <script src="/js/landing-page.js"></script>
+    <script src="js/helpers.js" type="module"></script>
+    <script src="js/fake-base-de-datos.js"></script>
+    <script src="js/landing-page.js"></script>
     <script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
     <script>
       AOS.init();
