+++ js/landing-page.js
@@ -1,4 +1,6 @@
-//SIMULAMOS BASE DE DATOS:
+// import { addFooter, colorScrollNav } from "./footer.js";
+
+//*SIMULAMOS BASE DE DATOS:
 class User{
   constructor(name, email, phoneNumber, gender, pass, admin, favs){
     this.name = name;
@@ -18,795 +20,15 @@ let users = [
   new User('Viviana Bacha', 'viviana@hotmail.com', '351-5678901',  'Femenino', '123456', true, []),
 ];
 
-let peliculas = [
-  {
-    "id": 1,
-    "nombre": "It",
-    "descripcion": "Remake del clásico de Stephen King en el que un payaso aterroriza a los niños de un vecindario. En un pequeño pueblo de Maine, siete niños conocidos como el Club de los Perdedores se encuentran cara a cara con problemas de la vida, matones y un monstruo que toma la forma de un payaso llamado Pennywise.",
-    "director": "Andy Muschietti",
-    "genero": "Terror, Fantasía",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 2017,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/6uDpMsdn8bbZScDptbika5JEtSH.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/tcheoA2nPATCm2vvXw2hVQoaEFD.jpg",
-    "video": "https://www.youtube.com/watch?v=xKJmEC5ieOk&ab_channel=WarnerBros.Pictures"
-  },
-  {
-    "id": 2,
-    "nombre": "Pulp Fiction",
-    "descripcion": "Jules y Vincent, dos asesinos a sueldo con no demasiadas luces, trabajan para el gángster Marsellus Wallace. Vincent le confiesa a Jules que Marsellus le ha pedido que cuide de Mia, su atractiva mujer. Jules le recomienda prudencia porque es muy peligroso sobrepasarse con la novia del jefe. Cuando llega la hora de trabajar, ambos deben ponerse manos a la obra . Su misión: recuperar un misterioso maletín",
-    "director": "Tarantino",
-    "genero": "Acción",
-    "categoria": "+18",
-    "publicada": false,
-    "destacada": true,
-    "año": "1994",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/fIE3lAGcZDV1G6XM5KmuWnNsPp1.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/tamCUHl16mptimLh1ZwJ1XH6QOr.jpg",
-    "video": "https://www.youtube.com/embed/tGpTpVyI_OQ"
-  },
-  {
-    "id": 3,
-    "nombre": "Todo Sobre Mi Madre",
-    "descripcion": "Madrid. Manuela, una madre soltera, ve morir a su hijo el día en que cumple 17 años, por echarse a correr para conseguir el autógrafo de Huma Rojo, su actriz favorita. Destrozada, Manuela viaja entonces a Barcelona en busca del padre del chico",
-    "director": "Almodobar",
-    "genero": "Drama",
-    "categoria": "+18",
-    "publicada": false,
-    "destacada": false,
-    "año": "1999",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/zmghiLIx11UZscUF3axGAUZ22Cn.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/u2CvJwAWDNk9w63gbOifmqd4W9r.jpg",
-    "video": "https://www.youtube.com/embed/jzMRQQmHBq4"
-  },
-  {
-    "id": 4,
-    "nombre": "Pacto de fuga",
-    "descripcion": "La noche del 29 de enero de 1990, poco antes de iniciarse la reconstrucción democrática en Chile, medio centenar de presos políticos se fugaron de la cárcel pública de Santiago por un túnel que 24 militantes del Frente Patriótico Manuel Rodríguez (FPMR) habían excavado durante 18 meses, 80 metros de longitud con herramientas tan rudimentarias como cucharas, tenedores y un solo destornillador; escondiendo 55 toneladas de tierra dentro del centro penitenciario",
-    "director": "David Albala",
-    "genero": "Acción",
-    "categoria": "+18",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/ziao1KWNkw4X5qL7rWCSJ6lyjX9.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/y1j02Kchix0fULJCrPZSzzR2qg9.jpg",
-    "video": "https://www.youtube.com/embed/qPx074pFbW8"
-  },
-  {
-    "id": 5,
-    "nombre": "El rey león",
-    "descripcion": "Simba es el hijo del rey de los leones, Mufasa, y heredero de todo el reino. Pero cuando su padre es brutalmente asesinado por su tío Scar, decidirá huir, dejando vía libre para que su tío tome el puesto de su padre como líder de la manada. En su camino, Simba se encuentra con el suricato Timón y el jabalí Pumba, que le enseñarán a vivir la vida sin preocupaciones. Pero el joven león se verá obligado a decidir entre su vida libre de problemas o su destino como rey.",
-    "director": "Jon Favreau",
-    "genero": "Familia, Animación",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2019,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/dzBtMocZuJbjLOXvrl4zGYigDzh.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/nRXO2SnOA75OsWhNhXstHB8ZmI3.jpg",
-    "video": "https://www.youtube.com/watch?v=TDIqwTgdT5U&ab_channel=DisneyEspa%C3%B1a"
-  },
-
-  {
-    "nombre": "Avatar: El sentido del agua",
-    "descripcion": "Ambientada más de una década después de los acontecimientos de la primera película, 'Avatar: The Way of Water' empieza contando la historia de la familia Sully (Jake, Neytiri y sus hijos), los problemas que los persiguen, lo que tienen que hacer para mantenerse a salvo, las batallas que libran para seguir con vida y las tragedias que sufren.",
-    "director": "James Cameron",
-    "genero": "Acción, Ciencia Ficción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "año": "2022",
-    "publicada": true,
-    "destacada": true,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/hP12za13SKPffvIVf8S59KzqSpZ.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/zENT3sNfL2HONYS99sFEYlE6R28.jpg",
-    "video": "https://www.youtube.com/watch?v=inT3MvpxRqE&ab_channel=20thCenturyStudiosEspa%C3%B1a",
-    "id": 6
-  }, 
-  {
-    "id": 7,
-    "nombre": "Big Fish",
-    "descripcion": "Cuenta la historia entre un padre moribundo, Edward Bloom (Albert Finney), y su hijo Will (Ewan McGregor), quien trata de aprender más acerca de su progenitor, reuniendo las distintas historias que este le ha contado. Así, Will recrea su vida en una serie de leyendas y mitos inspirados por los pocos hechos que conoce. Pero a partir del deseo de su madre Sandy, por reunir al padre con su hijo, este comienza a comprender la debacle de su padre. En sus relatos hay viajes alrededor del mundo, y delirios que incluyen gigantes, tornados y brujas.",
-    "director": "Tim Burton",
-    "genero": "Aventura, Fantasía, Drama",
-    "categoria": "+13",
-    "publicada": true,
-    "destacada": true,
-    "año": "2003",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/280BefgqWyLX0Z4jcbibB6l7xyx.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/zw2E53sYPSlBraYbrNnYRzcOaYr.jpg",
-    "video": "https://www.youtube.com/watch?v=R3AECE35Anw&t=3s&ab_channel=DisneyEspa%C3%B1a"
-  },
-  {
-    "nombre": "Bastardos sin gloria",
-    "descripcion": "Segunda Guerra Mundial. Durante la ocupación de Francia por los alemanes, Shosanna Dreyfus presencia la ejecución de su familia por orden del coronel nazi Hans Landa. Ella consigue huir a París, donde adopta una nueva identidad como propietaria de un cine. En otro lugar de Europa, el teniente Aldo Raine adiestra a un grupo de soldados judíos 'Los bastardos' para atacar objetivos concretos. Los hombres de Raine y una actriz alemana, que trabaja para los aliados, deben llevar a cabo una misión que hará caer a los jefes del Tercer Reich. El destino quiere que todos se encuentren bajo la marquesina de un cine donde Shosanna espera para vengarse.",
-    "director": "Quentin Tarantino",
-    "genero": "Acción, Bélica",
-    "categoria": "+13",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": true,
-    "año": "2009",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/2dUYqrrxA9Ar6LXGqaTftX286Jx.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/1Jpkm9qZcsT0mSyVXgs4VlGjPNI.jpg",
-    "video": "https://www.youtube.com/watch?v=PIoyKApfOsg&t=4s&ab_channel=20thCenturyStudiosEspa%C3%B1a",
-    "id": 8
-  },
-  {
-    "nombre": "Up",
-    "descripcion": "Tras atar miles de globos a su casa, Carl Fredricksen, un vendedor de globos jubilado, se embarca en el mundo de los sueños de la infancia. Pero sin que Carl lo sepa, Russell, un explorador de 8 años, se encuentra en el lugar equivocado en el momento equivocado: el porche delantero de la casa. El dúo más improbable conoce a fantásticos amigos como Dug, un perro con un collar especial que le permite hablar, y Kevin, un pájaro de cuatro metros de altura que no vuela. Atrapados en la selva, Carl se da cuenta de que a veces las mayores aventuras de la vida son las que menos esperabas.",
-    "director": "Pete Docter",
-    "genero": "Animación, Familia, Comedia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": true,
-    "año": "2009",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/zPlR7FOi4bqCZmPRvyox3FrNggc.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/hGGC9gKo7CFE3fW07RA587e5kol.jpg",
-    "video": "https://www.youtube.com/watch?v=qoZo9MiICo0&ab_channel=DisneyEspa%C3%B1al",
-    "id": 9
-  }, 
-  {
-    "id": 10,
-    "nombre": "El señor de los anillos: El retorno del rey",
-    "descripcion": "Las fuerzas de Saruman han sido destruidas, y su fortaleza sitiada. Ha llegado el momento de que se decida el destino de la Tierra Media, y por primera vez en mucho tiempo, parece que hay una pequeña esperanza. La atención del señor oscuro Sauron se centra ahora en Gondor, el último reducto de los hombres, y del cual Aragorn tendrá que reclamar el trono para ocupar su puesto de rey. Pero las fuerzas de Sauron ya se preparan para lanzar el último y definitivo ataque contra el reino de Gondor, la batalla que decidirá el destino de todos. Mientras tanto, Frodo y Sam continuan su camino hacia Mordor, a la espera de que Sauron no repare en que dos pequeños Hobbits se acercan cada día más al final de su camino, el Monte del Destino.",
-    "director": "Peter Jackson",
-    "genero": "Aventura, Fantasía, Acción",
-    "categoria": "+13",
-    "publicada": true,
-    "destacada": true,
-    "año": "2003",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/w6wTh6B6QYVozkFPqynEenMbPlk.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/rOmd5Beem8FNcEA2DrmlBMjf6v5.jpg",
-    "video": "https://www.youtube.com/watch?v=cUuz3-QCg58&ab_channel=ESmaugQDanVS"
-},
-  {
-    "id": 11,
-    "nombre": "El libro de la selva",
-    "descripcion": "Después de ser abandonado en la selva, una familia de lobos cría al pequeño cachorro humano Mowgli. Gracias a la manada de lobos, sale adelante hasta ser capaz de enfrentarse por sí mismo a los peligros de la selva. Pero con la llegada del peligroso Shere Khan, Mowgli deja de ser bien recibido. Este temible tigre, resentido por las heridas que le ha ocasionado el hombre, ha prometido eliminar toda amenaza humana en la selva. Así que a Mowgli no le queda más remedio que abandonar el único hogar que ha conocido en toda su vida. El pequeño niño humano emprenderá entonces un arriesgado y fascinante viaje de autodescubrimiento, para regresar junto a su especie, acompañado de la siempre fiel Bagheera, la pantera mentora del niño, y el despreocupado y divertido oso Baloo. Esta nueva adaptación de El Libro de la selva de Rudyard Kipling, que combina acción real con escenarios fotorrealistas y animales generados por ordenador, corre a cargo de Disney.",
-    "director": "Jon Favreau",
-    "genero": "Animación, Familia, Drama",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2016,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/pQGB08Kxyt3TLN4L01E175BRcvu.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/wcM0r4JcXK4Wz2LeP7qLU92ZOG9.jpg",
-    "video": "https://www.youtube.com/watch?v=C4qgAaxB_pc&ab_channel=WaltDisneyStudios"
-  },
-  {
-    "id": 12,
-    "nombre": "Raya y el último dragón",
-    "descripcion": "Raya, una niña de gran espíritu aventurero, se embarca en la búsqueda del último dragón del mundo con el objetivo de devolver el equilibrio a Kumandra, un lejano y recóndito territorio habitado por una civilización milenaria.",
-    "director": "Don Hall",
-    "genero": "Animación, Familia, Fantasía",
-    "categoria": "ATP",
-    "publicada": true,
-    "destacada": true,
-    "año": "2021",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/67ZrxjUEXPObtc2D5BDAjsXrnnR.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/79AoH5WVm4cWDZ20NGPhahLxNDX.jpg",
-    "video": "https://www.youtube.com/watch?v=R3AECE35Anw&t=3s&ab_channel=DisneyEspa%C3%B1a"
-  },
-  {
-    "id": 13,
-    "nombre": "Volver al futuro",
-    "descripcion": "Marty McFly, un estudiante de secundaria de 17 años, es enviado accidentalmente treinta años al pasado en un DeLorean que viaja en el tiempo, inventado por su gran amigo, el excéntrico científico Doc Brown.",
-    "director": "Robert Zemeckis",
-    "genero": "Aventura, Comedia, Ciencia Ficción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 1985,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/gcCbAMXptil3zInWAsZZ4ecH0Z9.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/ek6oBe077b5hDQ0iJlJh9XHRhVZ.jpg",
-    "video": "https://www.youtube.com/watch?v=U71BvFM7Wpw&ab_channel=VintageMovieTrailers"
-  },
-  {
-    "id": 14,
-    "nombre": "Matrix Revolutions",
-    "descripcion": "La tercera parte de “Matrix” forma junto a “Matrix Reloaded” una película dividida en dos debido a su larga duración. Matrix Revolutions tiene lugar en su mayor parte en en el mundo real, en el que Neo, Trinity y Morpheo lucharán contra el avance de las máquinas hacia la ciudad de Zion, último reducto de la raza humana.",
-    "director": "James McTeigue, Lana Wachowski, Lilly Wachowski",
-    "genero": "Acción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2003",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/oCB8tMvo5dd4j1NN2bLlXbgoeRB.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/9yJrf74rgOMfF5LrgUmLiJAEBSv.jpg",
-    "video": "https://www.youtube.com/embed/hMbexEPAOQI"
-  },
-  {
-    "id": 15,
-    "nombre": "A.I. Rising",
-    "descripcion": "Año 2048, Milutin ha emprendido un viaje a Alfa Centauri con la ayuda de Nimani, una androide con forma de mujer que, entre otras cosas, proporcionará compañía a Milutin, que la somete a una tórrida relación sexual. Harto de las funciones preestablecidas de la androide, el astronauta decide configurarla a partir de las experiencias que esta ha vivido… el resultado, sin embargo, no será el que él espera.",
-    "director": "Lazar Bodroža",
-    "genero": "Acción, Ciencia Ficción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2018",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/cvSjrRXBKzh5vGk01fUfrqSrkKO.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/kcsWiGL1HmtEeEpnR5zlCTt1RRS.jpg",
-    "video": "https://www.youtube.com/embed/iDdYMOiIn3Q"
-  },
-  {
-    "id": 16,
-    "nombre": "Bob Esponja: Un héroe fuera del agua",
-    "descripcion": "La vida en Fondo de Bikini no podría ser mejor para un eterno optimista como Bob Esponja (Tom Kenny) y sus amigos: la fiel estrella de mar Patricio (Bill Fagerbakke), el sarcástico Calamardo (Rodger Bumpass), la ardilla y científica Arenita (Carolyn Lawrence) y el crustáceo capitalista, Sr. Cangrejo (Clancy Brown). Cuando alguien roba la receta ultrasecreta de las Cangreburgers, los perpetuos enemigos Bob Esponja y Plankton (Mr. Lawrence) tendrán que unir sus fuerzas en un viaje a través del tiempo y el espacio para controlar sus superpoderes internos y enfrentarse al malvado pirata Barba Burger (Antonio Banderas), quien tiene sus propios planes para esas deliciosas exquisiteces.",
-    "director": "Bob Camp, Carla Rose Ponzio",
-    "genero": "Acción, Familia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2015",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/16XIGkJgoYwtecgF2OFavSnisZ7.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/c1ys6GmNTx8zWafOj2V4J8E2BGo.jpg",
-    "video": "https://www.youtube.com/embed/ISUMMHr1D3A"
-  },
-  {
-    "id": 17,
-    "nombre": "Los elegidos",
-    "descripcion": "Los Barret son una familia que vive en los suburbios. Lacey y Daniel (Keri Russell & Josh Hamilton) presencian una serie de extraños acontecimientos perturbadores que afectan a su familia, a su seguridad y a su tranquilidad. Cuando descubren que su familia es el objetivo de una inimaginable y terrorífica fuerza mortal, ambos tendrán que hacer todo lo que esté en sus manos para resolver este oscuro misterio.",
-    "director": "Jean-Paul Chreky, Renetta G. Amador, Scott Stewart",
-    "genero": "Suspense, Terror",
-    "categoria": "+16",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2013",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/mwqQ5QTJ0lDTm8Zd50vJZS5zGne.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/oLuAgHcYk9UrulfFJCmYfcnvmgB.jpg",
-    "video": "https://www.youtube.com/embed/7axk1-oDckA"
-  },
-  {
-    "id": 18,
-    "nombre": "Soul",
-    "descripcion": "¿Alguna vez te has preguntado de dónde provienen tu pasión, tus sueños y tus intereses? ¿Qué es lo que te hace ser... tú? Viaja desde las calles de Nueva York a los reinos cósmicos para descubrir las respuestas a las preguntas más importantes de la vida.",
-    "director": "Pete Docter",
-    "genero": "Comedia, Animación, Familia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2020,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/xJ7MCAy4oCLerPo3JjnB8lKOhl4.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/qF8eaVidzEsXGMU9qp8owby4bzV.jpg",
-    "video": "https://www.youtube.com/watch?v=3QIdlo4uIVg&ab_channel=SensaCineTRAILERS"
-  },
-  {
-    "id": 19,
-    "nombre": "Scarface",
-    "descripcion": "Cuenta la historia entre un padre moribundo, Edward Bloom (Albert Finney), y su hijo Will (Ewan McGregor), quien trata de aprender más acerca de su progenitor, reuniendo las distintas historias que este le ha contado. Así, Will recrea su vida en una serie de leyendas y mitos inspirados por los pocos hechos que conoce. Pero a partir del deseo de su madre Sandy, por reunir al padre con su hijo, este comienza a comprender la debacle de su padre. En sus relatos hay viajes alrededor del mundo, y delirios que incluyen gigantes, tornados y brujas.",
-    "director": "Brian De Palma",
-    "genero": "Acción, Crimen, Drama",
-    "categoria": "+16",
-    "publicada": true,
-    "destacada": false,
-    "año": "1983",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/oCd2HakBSfNU1J4LHY2nTvDs9W5.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/cCvp5Sni75agCtyJkNOMapORUQV.jpg",
-    "video": "https://www.youtube.com/watch?v=vSp_kLkNdnE&embeds_euri=https%3A%2F%2Fwww.themoviedb.org%2F&embeds_origin=https%3A%2F%2Fwww.themoviedb.org&source_ve_path=MjM4NTE&feature=emb_title"
-  },  
-  {
-    "nombre": "Pledge",
-    "descripcion": "Un grupo de estudiantes universitarios de primer año se comprometen a una fraternidad exclusiva, pero pronto se dan cuenta de que hay más en juego de lo que jamás hubieran imaginado.",
-    "director": "Daniel Robbins",
-    "genero": "Terror",
-    "categoria": "+13",
-    "codigo": "pls",
-    "año": "2019",
-    "publicada": true,
-    "destacada": false,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/3eol8FnndIY8drMknjUnRCjYaIe.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/aqdjM6chmFlOtt1bzW0uLVIJWan.jpg",
-    "video": "https://www.youtube.com/embed/o-fFgAQl0oM",
-    "id": 20
-  },
-  {
-    "id": 21,
-    "nombre": "The Purge 2: Anarchy",
-    "descripcion": "Secuela de la película The Purge. La noche de las bestias (2013). En la película original, el gobierno de Estados Unidos permite que, durante una noche, cualquier acto vandálico sea legal. De esta forma intentan reducir el número de criminales, ya que las cárceles del país se encuentran saturadas. La idea surge de la teoría de que todo ser humano tiene maldad en su interior, maldad que en algún momento saldrá a la luz. En esta situación, las personas que tienen dinero tienen más posibilidades que las pobres, ya que estas últimas no tienen los medios necesarios para defenderse de los que les quieran hacer daño. Así, una pregunta sale a la luz: ¿Cómo poder sobrevivir a los criminales sin convertirse en uno?",
-    "director": "Amy Price-Francis",
-    "genero": "Terror",
-    "categoria": "+18",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2014",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/lai6rkzGDOZk5CsPdKZAkzUhjlH.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/blfZnm6StMPnTY35kaooPBCsSVp.jpg",
-    "video": "https://www.youtube.com/embed/osI4Ii6a3D4"
-  },
-  {
-    "nombre": "Yo no soy tu mami",
-    "descripcion": "Una periodista sostiene en su columna semanal las razones por las cuales no quiere tener hijos, defendiendo su posición frente a la vida y la maternidad. Pero la misma comenzará a aflojarse cuando un nuevo vecino y su pequeña hija lleguen al edificio.",
-    "director": "Marcos Carnevale",
-    "genero": "Comedia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2019",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/5yRBIxOAnEaQOqy0eYHX1rJWGQh.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/gDeuJUEjrZTETMnYNGWubvrNmvT.jpg",
-    "video": "https://www.youtube.com/embed/cJVBYkA57yc",
-    "id": 22
-  },
-  {
-    "nombre": "Frozen Fever",
-    "descripcion": "Llega el cumpleaños de Anna, y Elsa y Kristoff están decididos a brindarle la mejor celebración de su vida, pero cuando Elsa se resfría, sus poderes helados pueden poner la fiesta – y algo más – en riesgo.",
-    "director": "Chris Buck, Jennifer Lee",
-    "genero": "Animación, Familia, Comedia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2015",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/2yvEHPlUky1b8nKaiWbI9k77Z6L.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/wD7s8gT9F89gIINlafDJNjYjzZ8.jpg",
-    "video": "https://www.youtube.com/embed/bQRLVxZHKPs",
-    "id": 23
-  },
-  {
-    "nombre": "3022",
-    "descripcion": "Un grupo de astronautas que viven en el inquietante vacío del espacio profundo despierta para encontrar la tierra ha sufrido un evento de nivel de extinción.",
-    "director": "John Suits",
-    "genero": "Ciencia Ficción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2019",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/qgdj1IjOs5ubfHaiexhuHNwPsYd.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/ailBwpIwzUcNpuKSHtSwyqXAYCe.jpg",
-    "video": "https://www.youtube.com/embed/zF3VsTjoQS0",
-    "id": 24
-  },
-  {
-    "id": 25,
-    "nombre": "El hombre invisible",
-    "descripcion": "Después de que su abusivo ex novio se suicide, la ingenua viuda Cecilia comienza a reconstruir su vida para mejor. Sin embargo, su sentido de la realidad se pone en tela de juicio cuando comienza a sospechar que su ex novio en realidad no está muerto, como había pensado.",
-    "director": "Leigh Whannell",
-    "genero": "Acción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/wCnsOikA3iab1MVE0zUjRtqz758.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/dYty1QmTwRANDfbm19Gu2PmNOEs.jpg",
-    "video": "https://www.youtube.com/embed/Yghnm_17a8w"
-  },
-  {
-    "nombre": "Avatar",
-    "descripcion": "Año 2154. Jake Sully , un exmarine en silla de ruedas, es enviado al planeta Pandora, donde se ha creado el programa Avatar, gracias al cual los seres humanos pueden controlar de forma remota un cuerpo biológico con apariencia y genética de la especie nativa. Pronto se encontrará con la encrucijada entre seguir las órdenes de sus superiores o defender al mundo que le ha acogido y siente como suyo.",
-    "director": "James Cameron",
-    "genero": "Acción, Ciencia Ficción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2009",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/loiA7RXohzfd6l0uoc22X7qQeCE.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/3BhYo9HUBlQ5XY3unbuqc2HaeYn.jpg",
-    "video": "https://www.youtube.com/watch?v=PIoyKApfOsg&t=4s&ab_channel=20thCenturyStudiosEspa%C3%B1a",
-    "id": 26
-  },
-  {
-    "nombre": "Starfish",
-    "descripcion": "Una chica intenta superar la pérdida de su mejor amiga justo cuando el mundo está apunto de desaparecer tal y como se ha conocido hasta el momento..",
-    "director": "AT White",
-    "genero": "Ciencia Ficción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2017",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/bn1uKQnkA32TjvLOvmYcQFe5NNe.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/kLksdhO1XzlweAOpRi3zVyNOexh.jpg",
-    "video": "https://www.youtube.com/embed/4inpM9tlock",
-    "id": 27
-  },
-  {
-    "nombre": "Incident at Montauk",
-    "descripcion": "Después de presenciar un OVNI, Jed Logan pronto descubre que su avistamiento puede estar relacionado con un proyecto secreto del gobierno que lo lleva por un camino peligroso para descubrir la verdad.",
-    "director": "Owen Mulligan",
-    "genero": "Ciencia Ficción",
-    "categoria": "+18",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2017",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/xW8IgkpKPICmBeLjBi4SXFu48Zb.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/3vZvgDdgdCWAbvELiA21uAFhu9z.jpg",
-    "video": "https://www.youtube.com/embed/bIBi8cdclHs",
-    "id": 28
-  },
-  {
-    "nombre": "Welcome to Sudden Death",
-    "descripcion": "Jesse Freeman is a former special forces officer and explosives expert now working a regular job as a security guard in a state-of-the-art basketball arena. Trouble erupts when a tech-savvy cadre of terrorists kidnap the team’s owner and Jesse’s daughter during opening night. Facing a ticking clock and impossible odds, it’s up to Jesse to not only save them but also a full house of fans in this highly charged action thriller.",
-    "director": "Dallas Jackson",
-    "genero": "Acción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/elZ6JCzSEvFOq4gNjNeZsnRFsvj.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/akflTTOROZVIyxvzzvCAOVChdbL.jpg",
-    "video": "https://www.youtube.com/embed/TpWiJWRSKIQ",
-    "id": 29
-  },
-  {
-    "nombre": "Primal",
-    "descripcion": "Un cazador que se dedica a capturar animales para luego venderlos a distintos zoológicos compra un pasaje para embarcarse en un crucero repleto de animales exóticos con los que piensa hacerse de oro, entre los que se incluye un jaguar y otros animales del Amazonas. Sin embargo, desconoce que en el mismo barco se está extraditando a un asesino en serie que, tras escapar del control de sus captores, liberta a los animales y con ellos el caos total.",
-    "director": "Nick Powell",
-    "genero": "Acción",
-    "categoria": "+18",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2019",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/v0Air5GTsfgtjsnZyji2lH6r2b8.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/5t7W5wuY3hlWjKRfgPIOWcWaDCL.jpg",
-    "video": "https://www.youtube.com/embed/xpxT6lWQDA4",
-    "id": 30
-  },
-  {
-    "nombre": "24 Hours In London",
-    "descripcion": "Un ex convicto vengativo realiza una matanza de 24 horas sin decir una palabra. El detective desacreditado que investiga el caso está convencido de que hay más de lo que sugiere la evidencia.",
-    "director": "Damien Morley, Dave Wiltshire, Paul Knight",
-    "genero": "Acción",
-    "categoria": "+18",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/3RrXYB2QilQ1xQXlRIncdVXTKIU.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/hlfFrHOCkdP7VgwRsJeeN5bJGRD.jpg",
-    "video": "https://www.youtube.com/embed/zKdbz4KNNAI",
-    "id": 31
-  },
-  {
-    "nombre": "Force of Nature",
-    "descripcion": "Una banda de ladrones planea un atraco durante un huracán y se encuentra con problemas cuando un policía en desgracia intenta obligar a todos en el edificio a evacuar.",
-    "director": "Michael Polish",
-    "genero": "Acción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/1qG8FzwGUlFMK3WCqPR0nJBwKUW.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/za4sjEoamjs1Ho7fGVpczjVYgTm.jpg",
-    "video": "https://www.youtube.com/embed/d61GX5VoEJc",
-    "id": 32
-  },
-  {
-    "nombre": "The Lie",
-    "descripcion": "Un padre y una hija se dirigen al campamento de baile cuando ven al mejor amigo de la niña al costado de la carretera. Cuando se detienen para ofrecerle un aventón al amigo, sus buenas intenciones pronto resultan en terribles consecuencias.",
-    "director": "Veena Sud",
-    "genero": "Drama",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2016",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/qjlpRjvnjDKwLJmK3Q2PPkj6PtU.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/tBJa5DSyDggtcbMSMNAKQQtpwBt.jpg",
-    "video": "https://www.youtube.com/embed/lrwhHRYDkgs",
-    "id": 33
-  },
-  {
-    "nombre": "The Last Boy",
-    "descripcion": "Con el fin del mundo, una madre moribunda envía a su hijo pequeño en una búsqueda para encontrar el lugar que concede deseos.",
-    "director": "Perry Bhandal",
-    "genero": "Drama",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2019",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/l1Cy0hcTDbjAntFfUneAUvazSiq.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/bNTBWrSLZrOdB4NLCV2LmA3fuiK.jpg",
-    "video": "https://www.youtube.com/embed/QB69KYPOKe8",
-    "id": 34
-  },
-  {
-    "nombre": "The Boys in the Band",
-    "descripcion": "Un grupo de homosexuales se reúne en un apartamento de Nueva York para celebrar el cumpleaños de un amigo. Cuando transcurren las horas, después de beber y de subir el volumen de la música, la velada comienza a exponer las fisuras que existen entre su amistad y el dolor auto-infligido que amenaza con hacer trizas su concepto de la solidaridad.",
-    "director": "Joe Mantello",
-    "genero": "Drama",
-    "categoria": "+16",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/2D2gi13hnrXWaspWuWmdbEZ8bef.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/oviYLyXIFKWiEdX8CXo0UD4tmUx.jpg",
-    "video": "https://www.youtube.com/embed/862Pb9oDDAo",
-    "id": 35
-  },
-  {
-    "nombre": "Clemency",
-    "descripcion": "Mientras se prepara para ejecutar a otro recluso, Bernadine debe enfrentarse a los demonios psicológicos y emocionales que crea su trabajo y, en última instancia, conectarla con el hombre al que está autorizada a matar.",
-    "director": "Chinonye Chukwu",
-    "genero": "Drama",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2019",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/uQgi5hq29saWiG4DbmQfz2EEnKI.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/17OdaqbR7JbavPrhKZsUWUSvqQz.jpg",
-    "video": "https://www.youtube.com/embed/HzUhz2XkFfE",
-    "id": 36
-  },
-  {
-    "nombre": "Carga",
-    "descripcion": "Un camión, una carretera y un destino impredecible, todos ellos intersecciones en una red de tráfico humano. Atrapada en esta red, Viktoriya solo tiene una opción: luchar para sobrevivir.",
-    "director": "Bruno Gascon",
-    "genero": "Drama",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2018",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/uLVjYX0q0wdX9CyILF6nMTRIP3x.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/ymePqzMInqJnyKBqul4x4SQpt68.jpg",
-    "video": "https://www.youtube.com/embed/IWyrDba0lDc",
-    "id": 37
-  },
-  {
-    "id": 38,
-    "nombre": "Interstellar",
-    "descripcion": "Narra las aventuras de un grupo de exploradores que hacen uso de un agujero de gusano recientemente descubierto para superar las limitaciones de los viajes espaciales tripulados y vencer las inmensas distancias que tiene un viaje interestelar.",
-    "director": "Christopher Nolan",
-    "genero": "Aventura, Drama, Ciencia Ficción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 2014,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/q6KI8ChlkIYYZww8C053cVfPQRo.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/w7F97hhhUPLVPWERoYGAiZrQpsz.jpg",
-    "video": "https://www.youtube.com/watch?v=OA3Txp94pjs&ab_channel=ApexClips"
-  },
-  {
-    "id": 39,
-    "nombre": "El señor de los anillos: Las dos torres",
-    "descripcion": "La Compañía del Anillo se ha disuelto. El portador del anillo Frodo y su fiel amigo Sam se dirigen hacia Mordor para destruir el Anillo Único y acabar con el poder de Sauron. Mientras, y tras la dura batalla contra los orcos donde cayó Boromir, el hombre Aragorn, el elfo Legolas y el enano Gimli intentan rescatar a los medianos Merry y Pipin, secuestrados por los ogros de Mordor. Por su parte, Saurón y el traidor Sarumán continúan con sus planes en Mordor, en espera de la guerra contra las razas libres de la Tierra Media.",
-    "director": "Peter Jackson",
-    "genero": "Aventura, Fantasía, Acción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2002,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/mTk8LNB3vEOojJlSrBKN2dDwF4g.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/nS4picOwj15APKzJeBCk6EBcMG5.jpg",
-    "video": "https://www.youtube.com/watch?v=nuTU5XcZTLA&ab_channel=WarnerBros.Entertainment"
-  },
-  {
-    "id": 40,
-    "nombre": "Perfume de mujer",
-    "descripcion": "Frank Slade es un malhumorado Coronel en la reserva del ejército norteamericano, retirado pues sufre de ceguera. Durante el fin de semana de Acción de Gracias el joven estudiante Charlie Simms, contratado por la familia de Slade, se queda en su casa para servirle de lazarillo y procurar que no beba mucho. Pero Frank tiene otros planes: irse a la gran ciudad de Nueva York... Remake de la película italiana de 1974 'Profumo di donna' (Perfume de mujer).",
-    "director": "Martin Brest",
-    "genero": "Drama",
-    "categoria": "+13",
-    "publicada": true,
-    "destacada": false,
-    "año": "1992",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/o4SBXaqUV6Xo11GE9xiQrRW3PtG.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/229egMytQoaZ79iRkVYuKKxmjJV.jpg",
-    "video": "https://www.youtube.com/watch?v=R3AECE35Anw&t=3s&ab_channel=DisneyEspa%C3%B1a"
-  },
-  {
-    "id": 41,
-    "nombre": "Pets united",
-    "descripcion": "Un astuto perro y una gatita consentida lideran a un equipo de héroes inesperados después de que su ciudad se vea asediada por el malvado alcalde y su ejército de robots.",
-    "director": "Reinhard Klooss",
-    "genero": "Comedia, Animación, Familia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 2019,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/wzJIhB6qrHJg75A0wGBiDeqS2mx.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/dp3jTcbb4KPOSOoItuUmp5ZvwEX.jpg",
-    "video": "https://www.youtube.com/embed/4Yq15X_gj9A"
-  },
-  {
-    "id": 42,
-    "nombre": "Mulán",
-    "descripcion": "Mulán, una joven china hija única de la familia Fa, en lugar de buscar novio, como sus amigas, trata por todos los medios de alistarse en el ejército imperial para evitar que su anciano padre sea llamado a filas para defender al Emperador del acoso de los Hunos. Cuando el emisario imperial lleva a cabo la orden de reclutar a los varones de todas las familias, Mulán se hará pasar por soldado y se someterá a un duro entrenamiento hasta hacerse merecedora de la estima y de la confianza del resto de su escuadrón.",
-    "director": "Barry Cook",
-    "genero": "Animación, Familia, Comedia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 1998,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/5wDwQcab7AcsM2zciw7gJmxUKEf.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/9I4LHCJRJ7j2M690r4Q3sK0VKyh.jpg",
-    "video": "https://www.youtube.com/watch?v=4XPc6bw6l9c&ab_channel=ReinaldoJunior"
-  },
-  {
-    "id": 43,
-    "nombre": "La boda de rosa",
-    "descripcion": "Rosa tiene un trabajo extenuante, un hermano mandón, un padre pegado como un mejillón, una hermana que no echa una mano, una hija que tras haberse emancipado vuelve a casa con dos bebés y un novio al que no consigue ver el pelo. Demasiado acostumbrada a poner a todo el mundo primero que a sí misma, Rosa se encuentra, a punto de cumplir cuarenta y cinco, no solo con que no tiene control sobre su vida, sino que por el camino que va, no lo va a tener nunca. Rosa decide dar un giro y tomar las riendas de su vida, o al menos intentarlo. Quiere cumplir el sueño de tener un negocio propio, pero antes, quiere casarse y embarcarse en un matrimonio muy especial.",
-    "director": "Icíar Bollaín",
-    "genero": "Comedia",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 2020,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/sxKwMF4rOyAaKiXcU5RfMyiIN8P.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/jp3LfaGGkXXS9Fq6361IzKoIy8v.jpg",
-    "video": "https://www.youtube.com/embed/jrAML3wM1ug"
-  },
-  {
-    "nombre": "El vuelo",
-    "descripcion": "Tras un aterrizaje de emergencia en medio del campo gracias al cual salvan la vida un centenar de pasajeros, el comandante Whip Whitaker, que pilotaba el avión, es considerado un héroe nacional. Sin embargo, cuando se pone en marcha la investigación para determinar las causas de la avería, se averigua que el capitán tenía exceso de alcohol en la sangre y que puede ir a la cárcel si se demuestra que pilotó el avión en estado de embriaguez.",
-    "director": "Robert Zemeckis",
-    "genero": "Drama",
-    "categoria": "+13",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": true,
-    "año": "2012",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/wc9INXYUM38pzpyZIgcCDhjv1a2.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/m9JmRJjqqh4SnHAEf7LQI8B07Kt.jpg",
-    "video": "https://www.youtube.com/watch?v=r-96m8c_o-E&ab_channel=ParamountPicturesInternational",
-    "id": 44
-  },
-  {
-    "id": 45,
-    "nombre": "Ted 2",
-    "descripcion": "Se trata de una secuela de la exitosa comedia escrita y dirigida por el cómico Seth MacFarlane. Recién casados Ted y Tami-Lynn deciden tener un bebé. John se ofrece en donar su esperma para que su mejor amigo pueda cumplir su sueño de tener un hijo, pero inesperadamente reciben una carta legal que dice que el estado no le permite ser padre ya que no está demostrado que sea una persona. Todos juntos tendrán que unir fuerzas para luchar por sus derechos en un tribunal de justicia.",
-    "director": "Kelly Cronin, Seth MacFarlane",
-    "genero": "Comedia",
-    "categoria": "A+13",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 2015,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/qMOt0uy1x49OBW0jzodgRM9waW0.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/kjOKnYJakcZ0sApcgcW3hBR8WPu.jpg",
-    "video": "https://www.youtube.com/embed/S3AVcCggRnU"
-  },
-  {
-    "nombre": "Books of Blood",
-    "descripcion": "Un viaje a territorios inexplorados y prohibidos a través de tres historias enredadas en el espacio y el tiempo.",
-    "director": "Brannon Braga",
-    "genero": "Terror",
-    "categoria": "ATP",
-    "año": "2020",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/bXLVZGgLgd7OwbfP6fC5GCfrxuD.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/a883xNT7wlw6GH38lvUkRRG1dIr.jpg",
-    "video": "https://www.youtube.com/embed/k-BVMDGFMkA",
-    "id": 46
-  },
-  {
-    "id": 47,
-    "nombre": "El señor de los anillos: La comunidad del anillo",
-    "descripcion": "En la Tierra Media, el Señor Oscuro Saurón creó los Grandes Anillos de Poder, forjados por los herreros Elfos. Tres para los reyes Elfos, siete para los Señores Enanos, y nueve para los Hombres Mortales. Secretamente, Saurón también forjó un anillo maestro, el Anillo Único, que contiene en sí el poder para esclavizar a toda la Tierra Media. Con la ayuda de un grupo de amigos y de valientes aliados, Frodo emprende un peligroso viaje con la misión de destruir el Anillo Único. Pero el Señor Oscuro Sauron, quien creara el Anillo, envía a sus servidores para perseguir al grupo. Si Sauron lograra recuperar el Anillo, sería el final de la Tierra Media.",
-    "director": "Peter Jackson",
-    "genero": "Aventura, Fantasía, Acción",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2002,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/w3NF110cyjqxkQD2azmUMyaFzhG.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/keyOZ8KcjHMTVH8p0owTXvw18Wf.jpg",
-    "video": "https://www.youtube.com/watch?v=FYZl0lpB9iM&ab_channel=IBroadcast"
-  },
-  {
-    "id": 48,
-    "nombre": "Se busca papá",
-    "descripcion": "Su mamá le prohibió inscribirse en una carrera de BMX, y esta pequeña cazadora de adrenalina no tiene mejor idea que contratar a un actor para el rol de su papá.",
-    "director": "Javier Colinas",
-    "genero": "Comedia",
-    "categoria": "ATP",
-    "publicada": false,
-    "destacada": false,
-    "año": "2020",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/pkGH2TfJpK9CnOB3qS8uVqAqdp9.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/fsDs4nGpA1w28mglVMu5LPgaeAj.jpg",
-    "video": "https://www.youtube.com/embed/UcUwCFSgGVQ"
-  },
-  {
-    "id": 49,
-    "nombre": "El sexto sentido",
-    "descripcion": "El doctor Malcom Crowe es un conocido psicólogo infantil de Philadelphia que vive obsesionado por el doloroso recuerdo de un joven paciente desequilibrado al que fue incapaz de ayudar. Cuando conoce a Cole Sear, un aterrorizado y confuso niño de ocho años que necesita tratamiento, ve que se le presenta la oportunidad de redimirse haciendo todo lo posible por ayudarlo. Sin embargo, el doctor Crowe no está preparado para conocer la terrible verdad acerca del don sobrenatural de su paciente: recibe visitas no deseadas de espíritus atormentados",
-    "director": "M. Night Shyamalan",
-    "genero": "Terror, Drama",
-    "categoria": "+13",
-    "publicada": true,
-    "destacada": false,
-    "año": "1999",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/vR8XCfaA2G3QdZmghuaXO5GdEYC.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/6TjllWT3cGrPFyqDXurVZ3L8bBi.jpg",
-    "video": "https://www.youtube.com/embed/cqJFb5ggwjk"
-  },
-  {
-    "id": 50,
-    "nombre": "Scream",
-    "descripcion": "Sidney Prescott, una adolescente de la pequeña comunidad de Woodsboro, se convierte en el objetivo de un misterioso asesino en serie. Una reportera de un canal de noticias, Gale Weathers, que cubrió el asesinato un año atrás de la madre de Sidney, cubre ahora estos asesinatos, además ha escrito un libro en el que defiende al supuesto asesino de Maureen, Cotton Weary, que va a ir a la cámara de gas, alegando que cayó en una trampa, lo que hace reavivar la polémica. Los crímenes dejan de ser aislados y continúan los asesinatos. Ahora Sidney, su novio y sus amigos están en peligro, ya que el asesino quiere revivir lo sucedido un año atrás con Maureen Prescott, pero esta vez con su hija.",
-    "director": "Wes Craven",
-    "genero": "Crimen, Terror, Misterio",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": false,
-    "publicada": true,
-    "año": 1996,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/ma59qEE7APF70Ekf06iSMAnUOA4.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/7ZigA0vEuUMDXeZdo8ltjov8oRB.jpg",
-    "video": "https://www.youtube.com/watch?v=23jmjs-rMGI&ab_channel=prinzefilms"
-  },  
-  {
-    "id": 51,
-    "nombre": "The Ring",
-    "descripcion": "Rachel Keller es una periodista de investigación que no da mucho crédito a lo que considera una especie de leyenda urbana: al parecer, circula por ahí una cinta de vídeo con imágenes aterradoras, que va acompañada de una llamada telefónica en la que se predice, con una semana de antelación, la muerte de quien las ha visto. Sin embargo, cuando cuatro adolescentes encuentran la muerte exactamente una semana después de ver el vídeo, la curiosidad se apodera de ella y no parará hasta encontrar la cinta y verla.",
-    "director": "Gore Verbinski",
-    "genero": "Terror, Misterio",
-    "categoria": "+13",
-    "codigo": "pls",
-    "destacada": true,
-    "publicada": true,
-    "año": 2002,
-    "tapa": "https://www.themoviedb.org/t/p/w1280/e2t5CKMox7tjv3iD3Ko7NdFa5lJ.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/9D5iDqq0BoitxyEsfnsHzBH0zzY.jpg",
-    "video": "https://www.youtube.com/watch?v=mTACCgSWu8E&ab_channel=RottenTomatoesClassicTrailers"
-  },
-  {
-    "nombre": "La caverna",
-    "descripcion": "Un grupo de estudiantes queda atrapado dentro de una cueva misteriosa donde descubren que el tiempo pasa de forma diferente bajo tierra que en la superficie.",
-    "director": "Ben Foster, Mark Dennis",
-    "genero": "Ciencia Ficción",
-    "categoria": "ATP",
-    "codigo": "pls",
-    "publicada": true,
-    "destacada": false,
-    "año": "2017",
-    "tapa": "https://www.themoviedb.org/t/p/w1280/2RV6aJqBztmkL0faL1aQsowtkPt.jpg",
-    "imagen": "https://www.themoviedb.org/t/p/original/lRFtFFIoGQfHxyHNS254Ew7PV9c.jpg",
-    "video": "https://www.youtube.com/watch?v=6T3Z0JTf77k&ab_channel=IndieWire",
-    "id": 26
-  },
-];
-
-if(!localStorage.getItem('films')){
-  let pelisJSON = JSON.stringify(peliculas)
-  localStorage.setItem('films', pelisJSON)
-};
-let filmsLS = localStorage.getItem('films')
-let dataFilms = JSON.parse(filmsLS);
-
 if(!localStorage.getItem('users')){
   let usersJSON = JSON.stringify(users);
   localStorage.setItem('users', usersJSON);
 };
+
+let filmsLS = localStorage.getItem('films')
+let dataFilms = JSON.parse(filmsLS);
 let usersLocalStorage = localStorage.getItem('users'); 
-console.log(usersLocalStorage);
 let dataUsersLS = JSON.parse(usersLocalStorage);
-console.log(dataUsersLS);
 
 function loginCheck(event){
   event.preventDefault();
@@ -815,12 +37,12 @@ function loginCheck(event){
   const userLoged = dataUsersLS.find(user=>user.email === email);
   console.log(email);
   if(userLoged){
-    console.log('email verificado');
+    console.log('LCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC');
     const passUserLoged = userLoged.pass === pass;
     console.log(pass);
     if (passUserLoged){
+    console.log('AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
       localStorage.setItem('user', JSON.stringify(userLoged));
-      localStorage.setItem("favs", JSON.stringify(userLoged.favs))
       window.location.assign(`${window.location.origin}/main.html`);
       console.log('Iniciaste sesión');
     }else{
@@ -1008,51 +230,6 @@ const addPreguntasFrecuentes =()=>{
 }
 addPreguntasFrecuentes()
 
-const addFooter =()=>{
-  let footer = document.createElement('div');
-  footer.classList.add('footer-landing', 'sticky-bottom')
-  footer.innerHTML = `
-          <div class="container-fluid">
-            <div class="row">
-              <div class="container-footer-links col-sm-6 col-md-4 order-1 mb-0 mb-sm-2 order-md-2 col-12">
-                <div>
-                  <div class="footer-links fs-5 my-2 py-2 my-0"><a class="text-decoration-none" href="/contact.html">Contacto</a></div>
-                  <div class="footer-links fs-5 my-2 py-2 my-0"><a class="text-decoration-none" href="/about-us.html">Sobre Nosotros</a></div>
-                  <div class="footer-links fs-5 my-2 py-2 my-0"><a class="text-decoration-none" href="/suport.html">Suport</a></div>
-                </div>
-              </div>
-              <div class="d-flex justify-content-center align-items-center my-3 col-md-4 order-0 order-md-0 col-12">
-                <a class="fw-bold fs-1 ms-xl-5 container-logo text-start" href="#header-landing">CineRolling+</a>
-              </div>
-              <div class="icons-footer col-12 col-sm-6 col-md-4 d-flex justify-content-evenly align-items-center order-2 order-md-1 ">
-                <div class="container-icons-redes d-flex d-flex justify-content-between my-4 my-sm-0">
-                  <div class="icons-redes footer-icon-insta text-light"><a class="icons-links-redes" href="/error404.html"><i class="fab fa-instagram"></i></a></div>
-                  <div class="icons-redes footer-icon-twitter text-light"><a class="icons-links-redes" href="/error404.html"><i class="fab fa-twitter"></i></a></div>
-                  <div class="icons-redes footer-icon-fb text-light"><a class="icons-links-redes" href="/error404.html"><i class="fab fa-facebook"></i></a></div>
-                </div>
-              </div>
-              <div class="container-footer-bottom order-4 text-light text-center col-12 d-flex justify-content-center py-2 mt-2 bg-dark">
-                <small>© 2023 <b>CineRolling</b> - Todos los Derechos Reservados.</small>
-              </div>
-            </div>
-          </div>
-  `
-  document.getElementsByTagName('footer')[0].appendChild(footer);
-}
-addFooter()
-
-const colorScrollNav =()=>{
-  let nav = document.querySelector('.navbar-landing');
-  window.addEventListener('scroll', function () {
-    if(window.pageYOffset>20){
-      nav.classList.add('bg-scroll-nav', 'shadow');
-    }else{
-      nav.classList.remove('bg-scroll-nav', 'shadow')
-    }
-  })
-}
-colorScrollNav()
-
 const rubberBandBtnSus =()=>{
   let buttonSuscript = document.querySelector('.button-landing-suscript');
   setInterval(() => {
@@ -1062,3 +239,4 @@ const rubberBandBtnSus =()=>{
     buttonSuscript.classList.remove('animate__rubberBand');
   }, 6000);
 }
+rubberBandBtnSus()
