+++ main.html
@@ -44,6 +44,8 @@
     <script>
       AOS.init();
     </script>
+    <!-- <script src="/js/landing-page.js" type="module"></script> -->
+    <script src="/js/helpers.js" type="module"></script>
     <script src="/js/main.js" type="module"></script>
   </body>
 </html>
