+++ movie-detail.html
@@ -24,10 +24,13 @@
 
     </header>
     <main>
+      <section class="img-movie-main d-flex align-items-end vh-100" aria-label="pelicula-principal">
+      </section>
     </main>
     <footer>
         
     </footer>
+    <script src="/js/helpers.js" type="module"></script>
     <script
     src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN"
